fix: disable project access tab when no access (#1057)
* refactor: remove unused import

* fix: disable project access tab when no access

* refactor: improve access denied text
